requirements.txt
anthropic>=0.40.0
openai>=1.50.0
notion-client>=2.2.1
feedparser>=6.0.11
requests>=2.32.0
python-dotenv>=1.0.0
click>=8.1.7
pydantic>=2.7.0
pydub>=0.25.1
